rename to schemaFile
